+++ foxes/models/model_book.py
@@ -402,6 +402,8 @@ class ModelBook:
             )
 
         self.wake_models[f"RHB"] = fm.wake_models.induction.RankineHalfBody()
+        self.wake_models[f"SelfSimilar"] = fm.wake_models.induction.SelfSimilar()
+        self.wake_models[f"SelfSimilar2020"] = fm.wake_models.induction.SelfSimilar2020()
 
         self.sources = Dict(
             name="sources",