+++ foxes/models/wake_models/induction/__init__.py
@@ -3,3 +3,5 @@ Induction wake models
 """
 
 from .rankine_half_body import RankineHalfBody
+from .self_similar import SelfSimilar
+from .self_similar2020 import SelfSimilar2020